Moved React rendering into message handler
