--- NavigationEdgeApi/Controllers/React.cs
@@ -1,28 +1,0 @@
-﻿using EdgeJs;
-using System;
-using System.Dynamic;
-using System.Threading.Tasks;
-
-namespace NavigationEdgeApi.Controllers
-{
-	public class React
-	{
-		private static Func<object, Task<object>> render = Edge.Func(@"
-			var React = require('react');
-			var Navigation = require('navigation');
-			var StateInfo = require('../../../node/StateInfo');
-				
-			StateInfo.register();
-			return function (data, callback) {
-				Navigation.StateController.navigateLink(data.url);
-				var component = React.createElement(Navigation.StateContext.state.component, data.props);
-				callback(null, React.renderToString(component));
-			}
-		");
-
-		public static Task<object> Render(string url, object props)
-		{
-			return render(new { url = url, props = props });
-		}
-	}
-}
