+++ NavigationEdgeApi/NavigationEdgeApi.csproj
@@ -95,7 +95,6 @@
   <ItemGroup>
     <Compile Include="App_Start\WebApiConfig.cs" />
     <Compile Include="Controllers\PeopleController.cs" />
-    <Compile Include="Controllers\React.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
